unit tests for handling cyclic references
